Moving 'git' and 'github' to kernel.
