+++ kernel/tools/git/utils.py
