+++ tests/kernel/intr/tools/git/test_gitpython.py
@@ -16,8 +16,8 @@
 from tempfile import TemporaryDirectory
 from unittest import TestCase
 
+from kernel.tools.git.gitpython import GitPython
 from tripleo.utils.fs import Dir
-from tripleo.utils.git.gitpython import GitPython
 from tripleo.utils.urls import URL
